fix(qfi and grantround): documentation upgrade and changes to GrantRound's `transferMatchingFunds`
Added missing documentation as well as upgraded certain descriptions to reflect the new changes of
the contracts. This documents the risks of using untrusted tokens and tokens with blacklisting
capabilities. Furthermore, the parameter `_erc20Address` within `transferMatchingFunds`
(`GrantRound`) was removed and the `nativeToken` is used instead. Furthermore, the contributor count
stored in the `QFI` contract is now reduced after a user withdraws their contribution.

fix #160, fix #161